+++ twilight-model/benches/deserialization.rs
@@ -2,12 +2,10 @@ use criterion::{criterion_group, criterion_main, Criterion};
 
 use serde::de::DeserializeSeed;
 use serde_json::Deserializer;
-use twilight_model::{
-    channel::Reaction,
-    gateway::{
-        event::GatewayEventDeserializer,
-        payload::incoming::{MemberChunk, TypingStart},
-    },
+use twilight_model::gateway::{
+    event::GatewayEventDeserializer,
+    payload::incoming::{MemberChunk, TypingStart},
+    GatewayReaction,
 };
 
 fn gateway_event_role_delete() {
@@ -151,7 +149,7 @@ fn reaction() {
         "user_id": "4"
     }"#;
 
-    serde_json::from_str::<Reaction>(input).unwrap();
+    serde_json::from_str::<GatewayReaction>(input).unwrap();
 }
 
 fn typing_start() {