+++ twilight-model/src/channel/mod.rs
@@ -1,4 +1,3 @@
-pub mod embed;
 pub mod message;
 pub mod permission_overwrite;
 pub mod stage_instance;
@@ -9,8 +8,6 @@ mod attachment;
 mod channel_mention;
 mod channel_type;
 mod followed_channel;
-mod reaction;
-mod reaction_type;
 mod video_quality_mode;
 
 use self::permission_overwrite::PermissionOverwrite;
@@ -20,8 +17,6 @@ pub use self::{
     channel_type::ChannelType,
     followed_channel::FollowedChannel,
     message::Message,
-    reaction::Reaction,
-    reaction_type::ReactionType,
     stage_instance::StageInstance,
     video_quality_mode::VideoQualityMode,
     webhook::{Webhook, WebhookType},